Codespaces merge into main (#993)
* Create devcontainer.json

* Adding feature for installing GMT directly

* Update devcontainer.json

* Update devcontainer.json - ubuntu 22.04

* Update devcontainer.json - just ubuntu

* cgroup providers will now check cgroupv1 paths if cgroupv2 paths are not found

* codespaces on ubuntu 22.04; gmt install on codespace container creation;  splash screeen; readme link;

* try install on post-create command to see if it helps with second run issue (etc/hosts gets deleted)

* rewrite to etc/hosts on start

* add api_url to on-start script

* add api_url to on-start script - wrong var name

* print_message -> echo

* make port changes for codespaces setup script

* set server_names_hash_bucket_size  in nginx.conf

* fix spacing in port setup

* fix port name for metrics, removed accidental paste

* added -t flag to codespace install script run

* add examples directory to codespace

* update codespace splash - note on port visibility

* add splash open to on-start script

* disable network io cgroup provider for codespace

* install reqs into codespace

* remove code splash screen line from codespace on-start

* add common category to disabled providers in codespace

* splash example is set to examples directory stress, not test stress

* try starting splash screen with postStartCommand

* remove splash start from postStartCommand - it couldn't find the right code application to open

* added containerized/base split; added workflow to build github codespace container; added devcontainer config for containerized version

* fix reference in contianerized devcontainer; removed duplicate docker install

* use secrets.github_token

* updated build and push action

* fix path for dockerfile

* try with checkout first, despite documentation

* use same base image

* remove trailing \ in mv command

* workspace -> workspaces

* corrected copy path for dockerfile

* dockerfile: cd into gmt directory then install

* turn off all but one psu providers

* install with no-hosts flag

* install docker compose

* activate venv before installing requirements

* add test image with nothing to find correct docker variables

* try new way of installing docker

* try last hail mary on installing from within dockerfile

* try one last thing

* try split containrization to avoid running docker in docker

* corrected context for gunicorn build

* try again with build context

* fix path?

* from instead of second pull

* don't need to build gunicorn again

* pull gunicorn image during codespace start, see how long that takes; run codespace-setup script on start

* remove test config; add codespace-startup script to containerized

* Speedy version for linux

* Added no ipmi and no sensors flag

* Added switches to entry statement

* Updated splash message

* Leveraging python cache even more

* Slimming down container

* Resureccting docker in docker and trimming splash screenh

* Removing docker in docker. Selfinstall

* Proper naming

* Splash fix

* Using codespaces default image

* Further savings on example clone

* New example applications folder

* Remote updated for XGboost

* Typo

* Moving to container build again

* Fixes for domain resolution in codespaces

* Deactiving GPU; Added true if directory exists; Stopping containers before new boot

* Updated splash message to use Bakery Demo Repo

* being more explicit about the Wagtail repo

* Updated idle to at least 10 s sleep

* Removed the no sleep to show container idle power

* Showing energy for every runtime phase

* Updated Cloud Energy

* Idle duration set to 10 to make container energy estimations more reliable

* Clarification on Turbo Boost and when it is off

* Improve Codespaces Setup (#833)

* Fix repo path in example

* minor

* Add VS Code extensions

* Add hint to make api port public to avoid CORS issues

* Remove YAML VS Code extension

* Open markdown files in preview mode by default

* Ignore backuped Makefile in Git

* Ensure ports are set to public after Codespace restart

* Update git repository URL

* Improve quickstart and add troubleshooting information

---------


* Disabling SSL for codespaces

* Loading block.conf also in Codespaces

* Updated XGBoost

* Typos in the container providers+

* Activating ports correctly + enabling XGboost manually

* Skipping lmsensors

* Refactored get cgroup path to C lib

* Compile errors and makefile additions

* Makefile additions

* Removing unused vars

* Using new --dev-cache-build switch. Was renamed

* Defautl host blocking must be disabled in GH Codespaces

* Check for memory file must be different

* Making codespaces work on main

---------

+++ lib/install_shared.sh
@@ -222,6 +222,9 @@ function build_binaries() {
             if [[ $(uname) == "Linux" ]] && [[ "$make_path" == *"/mach/"* ]]; then
                 continue
             fi
+            if [[ "$make_path" == *"/lmsensors/"* ]] && [[ "${install_sensors}" == false ]]; then
+                continue
+            fi
             echo "Installing $subdir/metric-provider-binary ..."
             rm -f $subdir/metric-provider-binary 2> /dev/null
             make -C $subdir
@@ -290,7 +293,6 @@ while getopts "p:a:m:nhtbisyrlc:k:" o; do
             ;;
         s)
             install_sensors=false
-            # currently unused
             ;;
         r)
             install_msr_tools=false
